Replaced ...client.mixins.json. Accidentally deleted it earlier and it broke the build.

+++ src/main/java/mod/netherrackfurnace/entity/NetherrackFurnaceBlockEntity.java
@@ -1,30 +1,16 @@
 package mod.netherrackfurnace.entity;
 
-import com.google.common.collect.Lists;
-import it.unimi.dsi.fastutil.objects.Object2IntMap;
-import it.unimi.dsi.fastutil.objects.ObjectIterator;
 import mod.netherrackfurnace.NetherrackFurnaces;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.AbstractFurnaceBlockEntity;
-import net.minecraft.block.entity.BlockEntityType;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.recipe.AbstractCookingRecipe;
-import net.minecraft.recipe.RecipeEntry;
 import net.minecraft.recipe.RecipeType;
 import net.minecraft.screen.FurnaceScreenHandler;
 import net.minecraft.screen.ScreenHandler;
-import net.minecraft.server.world.ServerWorld;
 import net.minecraft.text.Text;
-import net.minecraft.util.Identifier;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Vec3d;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.List;
-import java.util.Map;
 
 public class NetherrackFurnaceBlockEntity extends AbstractFurnaceBlockEntity {
 
@@ -46,7 +32,7 @@ public class NetherrackFurnaceBlockEntity extends AbstractFurnaceBlockEntity {
     protected int getFuelTime(ItemStack fuel) {
         if (!fuel.isEmpty()) {
             Item item = fuel.getItem();
-            if ((Integer) createFuelTimeMap().getOrDefault(item, 0) > 0) {
+            if (createFuelTimeMap().getOrDefault(item, 0) > 0) {
                 return Integer.MAX_VALUE;
             }
         } return 0;
